# paste-it-cleaned-app/swagger/swagger.json
{
    "swagger": "2.0",
    "info": { "version": "v1", "title": "PasteItCleaned Backend" },
    "paths": {
...
        "/analytics/hits": {
            "get": {
                "tags": ["Analytics"],
                "operationId": "GetAnalyticsHits",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [
                    { "name": "authorization", "in": "header", "required": false, "type": "string" },
                    { "name": "page", "in": "query", "required": false, "type": "integer", "format": "int32" },
                    { "name": "pageSize", "in": "query", "required": false, "type": "integer", "format": "int32" }
                ],
                "responses": { "200": { "description": "Success", "schema": { "$ref": "#/definitions/PagedListHitEntity" } }, "401": { "description": "Unauthorized" }, "403": { "description": "Forbidden" }, "500": { "description": "Server Error" } }
            }
        },
        "/analytics/hits/daily": {
            "get": {
                "tags": ["Analytics"],
                "operationId": "GetAnalyticsHitsDaily",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [
                    { "name": "authorization", "in": "header", "required": false, "type": "string" },
                    { "name": "startDate", "in": "query", "required": false, "type": "string", "format": "date-time" },
                    { "name": "endDate", "in": "query", "required": false, "type": "string", "format": "date-time" }
                ],
                "responses": {
                    "200": { "description": "Success", "schema": { "uniqueItems": false, "type": "array", "items": { "$ref": "#/definitions/HitDailyEntity" } } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "500": { "description": "Server Error" }
                }
            }
        },
        "/billing/invoices/{id}": {
            "get": {
                "tags": ["Billing"],
                "operationId": "GetInvoice",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [
                    { "name": "authorization", "in": "header", "required": false, "type": "string" },
                    { "name": "id", "in": "path", "required": true, "type": "string" }
                ],
                "responses": { "200": { "description": "Success", "schema": { "$ref": "#/definitions/InvoiceEntity" } }, "401": { "description": "Unauthorized" }, "403": { "description": "Forbidden" }, "500": { "description": "Server Error" } }
            }
        },
        "/billing/invoices": {
            "get": {
                "tags": ["Billing"],
                "operationId": "GetInvoices",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [{ "name": "authorization", "in": "header", "required": false, "type": "string" }],
                "responses": {
                    "200": { "description": "Success", "schema": { "uniqueItems": false, "type": "array", "items": { "$ref": "#/definitions/InvoiceEntity" } } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "500": { "description": "Server Error" }
                }
            }
        },
        "/billing/payment-method": {
            "get": {
                "tags": ["Billing"],
                "operationId": "GetPaymentMethod",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [{ "name": "authorization", "in": "header", "required": false, "type": "string" }],
                "responses": { "200": { "description": "Success", "schema": { "$ref": "#/definitions/PaymentMethodEntity" } }, "401": { "description": "Unauthorized" }, "403": { "description": "Forbidden" }, "500": { "description": "Server Error" } }
            },
            "post": {
                "tags": ["Billing"],
                "operationId": "PostPaymentMethod",
                "consumes": ["application/json-patch+json", "application/json", "text/json", "application/*+json"],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [
                    { "name": "authorization", "in": "header", "required": false, "type": "string" },
                    { "name": "obj", "in": "body", "required": false, "schema": { "$ref": "#/definitions/BillingRequest" } }
                ],
                "responses": { "200": { "description": "Success", "schema": { "$ref": "#/definitions/PaymentMethodEntity" } }, "401": { "description": "Unauthorized" }, "403": { "description": "Forbidden" }, "500": { "description": "Server Error" } }
            }
        },
        "/dashboard/hits": {
            "get": {
                "tags": ["Dashboard"],
                "operationId": "GetDashboardHits",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [{ "name": "authorization", "in": "header", "required": false, "type": "string" }],
                "responses": { "200": { "description": "Success", "schema": { "$ref": "#/definitions/PagedListHitEntity" } }, "401": { "description": "Unauthorized" }, "403": { "description": "Forbidden" }, "500": { "description": "Server Error" } }
            }
        },
        "/dashboard/hits/daily": {
            "get": {
                "tags": ["Dashboard"],
                "operationId": "GetDashboardHitsDaily",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [
                    { "name": "authorization", "in": "header", "required": false, "type": "string" },
                    { "name": "startDate", "in": "query", "required": false, "type": "string", "format": "date-time" },
                    { "name": "endDate", "in": "query", "required": false, "type": "string", "format": "date-time" }
                ],
                "responses": {
                    "200": { "description": "Success", "schema": { "uniqueItems": false, "type": "array", "items": { "$ref": "#/definitions/HitDailyEntity" } } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "500": { "description": "Server Error" }
                }
            }
        },
        "/dashboard/invoices": {
            "get": {
                "tags": ["Dashboard"],
                "operationId": "GetDashboardInvoices",
                "consumes": [],
                "produces": ["text/plain", "application/json", "text/json"],
                "parameters": [{ "name": "authorization", "in": "header", "required": false, "type": "string" }],
                "responses": {
                    "200": { "description": "Success", "schema": { "uniqueItems": false, "type": "array", "items": { "$ref": "#/definitions/InvoiceEntity" } } },
                    "401": { "description": "Unauthorized" },
                    "403": { "description": "Forbidden" },
                    "500": { "description": "Server Error" }
                }
            }
        },
...
    },
    "definitions": {
...
        "PagedListHitEntity": {
            "type": "object",
            "properties": {
                "results": { "uniqueItems": false, "type": "array", "items": { "$ref": "#/definitions/HitEntity" } },
                "currentPage": { "format": "int32", "type": "integer" },
                "pageCount": { "format": "int32", "type": "integer" },
                "pageSize": { "format": "int32", "type": "integer" },
                "rowCount": { "format": "int32", "type": "integer" },
                "firstRowOnPage": { "format": "int32", "type": "integer", "readOnly": true },
                "lastRowOnPage": { "format": "int32", "type": "integer", "readOnly": true }
            }
        },
        "HitEntity": {
            "type": "object",
            "properties": {
                "hitId": { "format": "uuid", "type": "string" },
                "date": { "format": "date-time", "type": "string" },
                "ip": { "type": "string" },
                "price": { "format": "double", "type": "number" },
                "type": { "type": "string" },
                "referer": { "type": "string" }
            }
        },
        "HitDailyEntity": {
            "type": "object",
            "properties": {
                "hitDailyId": { "format": "uuid", "type": "string" },
                "date": { "format": "date-time", "type": "string" },
                "countExcel": { "format": "int32", "type": "integer" },
                "countWord": { "format": "int32", "type": "integer" },
                "countPowerPoint": { "format": "int32", "type": "integer" },
                "countWeb": { "format": "int32", "type": "integer" },
                "countText": { "format": "int32", "type": "integer" },
                "countImage": { "format": "int32", "type": "integer" },
                "countOther": { "format": "int32", "type": "integer" },
                "totalPrice": { "format": "double", "type": "number" }
            }
        },
        "InvoiceEntity": {
            "type": "object",
            "properties": {
                "invoiceId": { "format": "uuid", "type": "string" },
                "number": { "format": "int32", "type": "integer" },
                "date": { "format": "date-time", "type": "string" },
                "price": { "format": "double", "type": "number" },
                "taxes": { "format": "double", "type": "number" },
                "total": { "format": "double", "type": "number" },
                "paid": { "type": "boolean" },
                "paidOn": { "format": "date-time", "type": "string" },
                "payment": { "$ref": "#/definitions/PaymentEntity" }
            }
        },
        "PaymentEntity": {
            "type": "object",
            "properties": { "paymentId": { "format": "uuid", "type": "string" }, "trxNumber": { "type": "string" }, "total": { "format": "double", "type": "number" }, "paymentMethod": { "$ref": "#/definitions/PaymentMethodEntity" } }
        },
        "PaymentMethodEntity": { "type": "object", "properties": { "paymentMethodId": { "format": "uuid", "type": "string" }, "type": { "type": "string" }, "paymentFailed": { "type": "boolean" } } },
        "BillingRequest": { "type": "object", "properties": { "any": { "type": "string" } } },
...
    }
}